% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphMain.R
\name{addEventLines}
\alias{addEventLines}
\title{add lines to dygrap plot representing begin and end of events}
\usage{
addEventLines(
  dygraph,
  events,
  labels = seq_len(nrow(events)),
  color = "red",
  labelLoc = "top",
  color.begin = color,
  color.end = color,
  labelLoc.begin = labelLoc,
  labelLoc.end = labelLoc,
  signalWidth = kwb.event::hsSigWidth(events)
)
}
\arguments{
\item{dygraph}{dygraph object as returned by \link[dygraphs]{dygraph}}

\item{events}{data frame with columns \emph{tBeg} (begin time) and \emph{tEnd}
(end time) of events to be drawn}

\item{labels}{labels given to the events (default: \code{1:nrow(events)})}

\item{color}{color (default: "red")}

\item{labelLoc}{"top" or "bottom" (default: "top")}

\item{color.begin}{default: same as argument "color"}

\item{color.end}{default: same as argument "color"}

\item{labelLoc.begin}{default: same as argument "labelLoc"}

\item{labelLoc.end}{default: same as argument "labelLoc"}

\item{signalWidth}{signal width in seconds = length of time interval that
one timestamp in the original data represents, e.g. 300 for 5-minutes-data.}
}
\value{
modified dygraph object
}
\description{
add lines to dygrap plot representing begin and end of events
}
