% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphMain.R
\name{addEventShades}
\alias{addEventShades}
\title{add shaded areas representing time intervals within "events"}
\usage{
addEventShades(
  dygraph,
  events,
  color = "lightgrey",
  signalWidth = kwb.event::hsSigWidth(events)
)
}
\arguments{
\item{dygraph}{dygraph object as returned by \link[dygraphs]{dygraph}}

\item{events}{data frame with columns \emph{tBeg} (begin time) and \emph{tEnd}
(end time) of events to be drawn}

\item{color}{color (default: "lightgrey")}

\item{signalWidth}{signal width in seconds = length of time interval that one
timestamp in the original data represents, e.g. 300 for 5-minutes-data}
}
\value{
modified dygraph object
}
\description{
add shaded areas representing time intervals within "events"
}
