% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphMain.R
\name{dyEvents}
\alias{dyEvents}
\title{call dygraphs::dyEvent repeatedly}
\usage{
dyEvents(
  dygraph,
  dates,
  labels = "",
  labelLocs = c("top", "bottom"),
  colors = "black",
  strokePatterns = "dashed"
)
}
\arguments{
\item{dygraph}{Dygraph to add event line to}

\item{dates}{vector of Date/time for the event (must be a POSIXct object or
another object convertible to POSIXct via as.POSIXct)}

\item{labels}{vector of labels for the events. (default: "")}

\item{labelLocs}{vector of locations for the labels ("top" or "bottom").
default: c("top", "bottom")}

\item{colors}{vector of colors for each event line. This can be of the
form "#AABBCC" or "rgb(255,100,200)" or "yellow". Defaults to black.}

\item{strokePatterns}{vector of predefined stroke pattern types ("dotted",
"dashed", or "dotdash")}
}
\value{
modified dygraph object
}
\description{
call dygraphs::dyEvent repeatedly
}
